Allow customizable path and filename for .htmlhintrc
This commit allows a user to specify a custom file path and file name
for a .htmlhintrc file. This is useful in cases where someone wants to
  keep a versioned dotfiles repository, but does not want to worry about
  symlinks on a machine. The syntax in the config.cson file of atom is:

  linter:
    "linter-htmlhint":
       htmlhintRcFilePath: "/Users/jacobsmith/"
       htmlhintRcFileName: ".htmlhintrc.custom"

Obviously, these paths and filenames can be any valid path and
configuration file for htmlhint. Please note that the htmlhintrc
fileName defaults to '.htmlhintrc' so it does not need overridden if
that name is not changing. Zero, one, or both of the options can be set.

Also, just for clarification's sake, the
.htmlhintrc file is the configuration file of the actual binary htmlhint
(https://github.com/yaniswang/HTMLHint), not linter-htmllint.