Add config watcher for no .htmlhintrc, don't lint without config
Add a configuration watcher for the new disableWhenNoHtmhintConfig
options (same as in linter-eslint). And prevent the linter from doing
anything if no ruleset is found and the disableWhenNoHtmlhintConfig is
set to `true`.

Also added top level divider comments to lib/index.js, similar to
src/main.js over at linter-eslint.