feat: disable when no .htmlhintrc file is found (#172)
* Add config option to disable linter-htmlhint when not .htmlhintrc found

The formattig is taken from the linter-eslint
(https://github.com/AtomLinter/linter-eslint) package.

* Add config watcher for no .htmlhintrc, don't lint without config

Add a configuration watcher for the new disableWhenNoHtmhintConfig
options (same as in linter-eslint). And prevent the linter from doing
anything if no ruleset is found and the disableWhenNoHtmlhintConfig is
set to `true`.

Also added top level divider comments to lib/index.js, similar to
src/main.js over at linter-eslint.

* Add unit tests for disable option

Added two unit tests for the new 'disable when no .htmlhintrc file is
found' options. One to see if a default configuration is used when no
.htmlhintrc file is present and the option IS NOT set, and one to see if
the file is not linted if the option IS set.